Fix MapServer/x : Add LayerRoot filter handling and fix return null
So I don't understand everything but I have something that look to works.

RestHandlerOpCode.LayerRoot is not managed but existing within the system so I've added it for correct management.

On top it seem to me that calling /Mapserver/1 is triggering an "internal" call that is made by the server, and has no user, so no rights. Returning null in the root request sample seem to break this internal call.
Filtering of the layers is done anyway in post filter operation that seem to works as expected.